Разбил крупные файлы на более мелкие по типам.
