+++ internal/models/accrual.go
@@ -1,0 +1,7 @@
+package models
+
+type AccrualResponse struct {
+	Number  string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual"`
+}
