+++ internal/models/balance.go
@@ -1,0 +1,6 @@
+package models
+
+type UserBalance struct {
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
+}
