+++ internal/models/order.go
@@ -1,0 +1,18 @@
+package models
+
+import "time"
+
+type OrderData struct {
+	Number     int64     `json:"number"`
+	UserID     string    `json:"user_id"`
+	Status     string    `json:"status"`
+	Accrual    float64   `json:"accrual"`
+	UploadedAt time.Time `json:"uploaded_at"`
+}
+
+type GetUserOrdersResponse struct {
+	Number     string    `json:"number"`
+	Status     string    `json:"status"`
+	Accrual    float64   `json:"accrual"`
+	UploadedAt time.Time `json:"uploaded_at"`
+}
