+++ internal/models/user.go
@@ -1,0 +1,7 @@
+package models
+
+type UserData struct {
+	UserID   string `json:"user_id"`
+	Login    string `json:"login"`
+	Password string `json:"password"`
+}
