+++ internal/storage/db_errors/db_errors.go
@@ -1,0 +1,12 @@
+package db_errors
+
+import (
+	"errors"
+
+	_ "github.com/jackc/pgx/v5/stdlib"
+)
+
+var ErrUserOrderDuplicate = errors.New("current user data conflict")
+var ErrOrderDuplicate = errors.New("data conflict")
+var ErrNotFound = errors.New("data not found")
+var ErrUserNoBalance = errors.New("user have not balance")
