+++ internal/storage/init_storage.go
@@ -2,11 +2,12 @@ package storage
 
 import (
 	"github.com/nu-kotov/gophermart/internal/config"
+	"github.com/nu-kotov/gophermart/internal/storage/postgres"
 )
 
-func NewStorage(c config.Config) (*DBStorage, error) {
+func NewStorage(c *config.Config) (*postgres.DBStorage, error) {
 
-	DBStorage, err := NewConnect(c.DatabaseConnection)
+	DBStorage, err := postgres.NewConnect(c.DatabaseConnection)
 	if err != nil {
 		return nil, err
 	}
